backend/requirements.txt
# Base
pytz==2021.1
psycopg2==2.8.6 --no-binary psycopg2
Pillow==8.1.2

# Django
django==3.2.4

# Django REST Framework
djangorestframework==3.12.2
django-filter==2.4.0

# JWT
pyjwt==2.1.0

# Debugging
ipdb==0.13.9